Some files listed as to be deleted (according to icon in staged changes area). 0010-Extract-Goal.sql 0010-Extract-SpedStaffMember.sql Transform_IepService.sql Transform_PrgGoals.sql Transform_Schedule.sql Transform_ServiceProviderTitle.sql
To be updated include:
0002-LoadMaps.sql
Transform_Section.sql

To be added:
0000-RegisterModule.sql

I'm not sure how these appeared in the Staged Changes area, because they still appear in the Unstaged Changes area.
I suspect that they show in the staged changes are because of a commit/push that failed because of a non-fast-forward error.

I have run Remote: Fetch twice.  The first result, I believe, was "success", but my attempt to push ammeded commit failed again.
The second result looks different because of the files listed here, which may be because this time I selected "New Commit" instead of "Amend Last Commit"